+++ actix-web-lab/src/lib.rs
@@ -44,6 +44,7 @@ mod local_data;
 mod middleware_from_fn;
 mod ndjson;
 mod normalize_path;
+mod panic_reporter;
 mod path;
 mod query;
 mod redirect;