+++ actix-web-lab/src/middleware.rs
@@ -6,5 +6,6 @@ pub use crate::catch_panic::CatchPanic;
 pub use crate::err_handler::ErrorHandlers;
 pub use crate::middleware_from_fn::{from_fn, MiddlewareFn, Next};
 pub use crate::normalize_path::NormalizePath;
+pub use crate::panic_reporter::PanicReporter;
 pub use crate::redirect_to_https::RedirectHttps;
 pub use crate::redirect_to_www::redirect_to_www;